ingestion/requirements.txt
psycopg2-binary
openai
python-dotenv
beautifulsoup4
pgvector
